app/requirements.txt
pandas
streamlit
sqlalchemy
python_dotenv
psycopg2-binary
